indexer/requirements.txt
elastic-transport==8.4.0
elasticsearch==8.6.2
pandas==1.5.3
python-dotenv==1.0.0
retry==0.9.2
